fix: redemption rate bug in reconfiguration (#4107)
## What does this PR do and why?

Fixes a bug in the reconfiguration settings where when a user disables the redemption rate, it sets it to 100; instead it now sets it to 0.

I have also removed disabling the save changes button when there is no changes as it is flaky, and may not behave as user expects. As the user can open this modal and then get feedback whether they can actually deploy or not, I think it is okay for now.

This also fixes the ballot redemption rate, by automatically setting it the same as the redemption rate.

## Steps to Repeat Initial Bug

1. Create a project with redemption rate set. For ease of testing, set to 0 < x < 100
1. Open settings and disable redemption rate with switch
1. Deploy
1. See error (redemption rate will be incorrectly set to 100)

It is **expected** that the redemption rate be set to 0.

## Screenshots or screen recordings

_If applicable, provide screenshots or screen recordings to demonstrate the changes._

## Acceptance checklist

- [ ] I have evaluated the [Approval Guidelines](https://github.com/jbx-protocol/juice-interface/blob/main/CONTRIBUTING.md#approval-guidelines) for this PR.
- [ ] (if relevant) I have tested this PR in [all supported browsers](https://github.com/jbx-protocol/juice-interface/blob/main/CONTRIBUTING.md#supported-browsers).
- [ ] (if relevant) I have tested this PR in dark mode and light mode (if applicable).